requirements.txt
pyTelegramBotAPI
python-dotenv
paypalrestsdk
python-binance
requests
Flask

pytest
